Added the trial interop to the HTML template
I added a comment to the HTML template for the JS interop. This needs to be tested to see if it displays properly once rendered on the docs page. I made this change because it was easy to overlook the different JS paths when doing setup. If the Trial path was inside of an HTML block it would be harder to miss.