Update grid-conditional-expand-button.md
Setting the CSS class "no-children" was inverted.

I was having a "null" issue because the row didn't have childs. The error was style showing if I clicked on the cell after hidding the expand icon.
To fix this, I added a CSS class to set the visibility of the .k-hierarchy-cell to "hidden" so the click is not triggered when clicking in the cell.