Update command.md (#1104)
Updated the exemple code to be consistent with the rest of the doc. 
I mainly saw that function receiving a xxxxEventArgs parameter, the property is "args" and not "e". 
In the rest of the doc, the function name for events is in the format of xxxx{Event}Handler. Exemple : xxxxOnClickHandler, xxxxOnSaveHandler.
I also added "private" in front of some function that didn't add an accessor defined, just to make sense with the function that had one.